\name{EPAM110}
\docType{data}
\alias{EBLOSUM50}
\alias{EBLOSUM62}
\alias{EPAM110}
\title{Amino Acid Substitution matrices}
\description{
Amino Acid Substitution matrices
}
\usage{data(EPAM110)}
\format{}
\details{
Amino Acid Substitution Matrices.
}
\source{
\url{http://www.hgmp.mrc.ac.uk/Software/EMBOSS/}
}
\references{http://r4protoemics.sourceforge.net}
\examples{
data(EPAM110)
}
\keyword{datasets}
