\name{raw}
\alias{raw}
\docType{data}
\title{marrayRaw example data}
\description{
  An object of class \code{marrayRaw} containing the
  raw data of a dual channel microarray experiment.
}
\usage{data(raw)}
\format{
  An object of class \code{marrayRaw}
}
\references{
Uva, P., and de Rinaldis, E. CrosshybDetector: detection of cross-hybridization
events in DNA microarray experiments. Submitted.
}
\examples{
data(raw)
str(raw)
}
\keyword{datasets}
